Add toolbarActions slot to DashboardLayout
