+++ packages/toolpad-core/src/Account/Account.tsx
@@ -7,7 +7,7 @@ import Popover from '@mui/material/Popover';
 import Stack from '@mui/material/Stack';
 import Typography from '@mui/material/Typography';
 import LogoutIcon from '@mui/icons-material/Logout';
-import { SessionContext, AuthenticationContext } from '../AppProvider/AppProvider';
+import { SessionContext, AuthenticationContext } from '../shared/context';
 import { SessionAvatar } from './SessionAvatar';
 
 export interface AccountProps {
