+++ packages/toolpad-core/src/AppProvider/AppProvider.tsx
@@ -8,6 +8,8 @@ import {
   BrandingContext,
   NavigationContext,
   RouterContext,
+  AuthenticationContext,
+  SessionContext,
   WindowContext,
 } from '../shared/context';
 import { AppThemeProvider } from './AppThemeProvider';
@@ -69,10 +71,6 @@ export interface Authentication {
   signIn: () => void;
   signOut: () => void;
 }
-
-export const SessionContext = React.createContext<Session | null>(null);
-
-export const AuthenticationContext = React.createContext<Authentication | null>(null);
 
 export type AppTheme = Theme | { light: Theme; dark: Theme };
